requirements.txt
django
pillow
psycopg2-binary
requests
discord.py
